dnscache: introduce global and per thread init/deinit functions
This patch splits the DNS cache initialization functions to a global
init/deinit to be called at startup time and a per-thread init/deinit
to be called at thread startup/shutdown time.

The DNS cache API should be refactored in a future point as it uses
a number of global variables where it should be using a GlobalConfig
specific state.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>